Specify commons-text version with correct JaroWinklerDistance implementation.
Version 1.10.0 of Apache artifact commons-text correctly calculates the JaroWinkler "distance". Version 1.9 was actually buggy and returning the "similarity", which is (1.0-distance).

The correct implementation is important, since it is used by the apoc.text.jaroWinklerDistance function. And if a custom plugin has a function or procedure dependent on apoc.text.jaroWinklerDistance, then its integration test can fail with the wrong version of commons-text.